+++ R/wq.R
@@ -13,8 +13,9 @@
 #'   "SLS" (Smelt Larva Survey, \code{\link{SLS}}),
 #'   "Baystudy" (Bay Study, \code{\link{baystudy}}),
 #'   "USGS" (USGS San Francisco Bay Surveys, \code{\link{USGS}}),
-#'   "USBR" (United States Bureau of Reclamation Sacramento Deepwater Ship Channel data, \code{\link{USBR}}), and
-#'   "Suisun" (Suisun Marsh Fish Study, \code{\link{suisun}}).
+#'   "USBR" (United States Bureau of Reclamation Sacramento Deepwater Ship Channel data, \code{\link{USBR}}),
+#'   "Suisun" (Suisun Marsh Fish Study, \code{\link{suisun}}), and
+#'   "YBFMP" (Yolo Bypass Fish Monitoring Program, \code{\link{YBFMP}}).
 #' @param Start_year Earliest year you would like included in the dataset. Must be an integer. Defaults to year \code{0}.
 #' @param End_year Latest year you would like included in the dataset. Must be an integer. Defaults to the current year.
 #' @importFrom magrittr %>%
@@ -23,7 +24,7 @@
 #' @examples
 #' Data <- wq(Sources = c("EMP", "STN", "FMWT", "EDSM",
 #' "DJFMP", "SDO", "SKT", "SLS",
-#' "20mm", "Suisun", "Baystudy", "USBR", "USGS"))
+#' "20mm", "Suisun", "Baystudy", "USBR", "USGS", "YBFMP"))
 #' @export
 
 wq <- function(Sources=NULL,
@@ -34,9 +35,9 @@ wq <- function(Sources=NULL,
 # Check arguments ---------------------------------------------------------
 
 if(is.null(Sources) | !all(Sources%in%c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SDO", "SKT", "SLS",
-                                        "20mm", "Suisun", "Baystudy", "USBR", "USGS"))){
+                                        "20mm", "Suisun", "Baystudy", "USBR", "USGS", "YBFMP"))){
   stop('You must specify the data sources you wish to include. Choices include
-  c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SDO", "SKT", "SLS", "20mm", "Suisun", "Baystudy", "USBR", "USGS")')
+  c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SDO", "SKT", "SLS", "20mm", "Suisun", "Baystudy", "USBR", "USGS", "YBFMP)')
 }
 
   # Set end year to current year if blank
@@ -102,6 +103,10 @@ if(is.null(Sources) | !all(Sources%in%c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "
 
   if("USGS"%in%Sources){
     WQ_list[["USGS"]]<-discretewq::USGS
+  }
+
+  if("YBFMP"%in%Sources){
+    WQ_list[["YBFMP"]]<-discretewq::YBFMP
   }
 
   out<-dplyr::bind_rows(WQ_list)%>%