+++ data-raw/SDO.R
@@ -18,16 +18,17 @@ SDO_stations<-read_csv(file.path(tempdir(), "IEP_DOSDWSC_site_locations_latitude
 SDO<-read_csv(file.path(tempdir(), "IEP_DOSDWSC_1997_2018.csv"),
               col_types=cols_only(Date="c", Time="c", StationID="c",
                                   WTSurface="d", WTBottom="d", SpCndSurface="d",
-                                  Secchi="d"#, Microcystis="c"
-                                  ))%>%
+                                  Secchi="d", Microcystis="c"))%>%
   left_join(SDO_stations, by="StationID")%>%
   rename(Station=StationID, Temperature=WTSurface, Temperature_bottom=WTBottom,
          Conductivity=SpCndSurface)%>%
   mutate(Source="SDO",
          Date=parse_date_time(Date, orders="%m/%d/%Y", tz="America/Los_Angeles"),
          Time=stringr::str_pad(Time, width=4, side="left", pad="0"),
-         Datetime=parse_date_time(if_else(is.na(Time), NA_character_, paste(Date, Time)), orders="%Y-%m-%d %H%M", tz="America/Los_Angeles"))%>%
-  select(Source, Station, Latitude, Longitude, Date, Datetime, Secchi, Temperature, Temperature_bottom, Conductivity)
+         Datetime=parse_date_time(if_else(is.na(Time), NA_character_, paste(Date, Time)), orders="%Y-%m-%d %H%M", tz="America/Los_Angeles"),
+         Microcystis=recode(Microcystis, present=NA_character_, absent=NA_character_),
+         Microcystis=round(as.numeric(Microcystis)))%>%
+  select(Source, Station, Latitude, Longitude, Date, Datetime, Microcystis, Secchi, Temperature, Temperature_bottom, Conductivity)
 
 
 usethis::use_data(SDO, overwrite = TRUE)