+++ NEWS.md
@@ -1,3 +1,4 @@
+* Added the Yolo Bypass Fish Monitoring Program data
 * Added Stockton Dissolved Oxygen survey as an additional data source
 * Removed duplicated water quality data from suisun data
 * Added Smelt Larva Survey as an additional data source